+++ src/navigation/rootStackRoutes.ts
@@ -16,8 +16,6 @@ You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { PossibleNextScreenAfterDestination } from '~/navigation/SendNavigation'
-
 type RootStackParamList = {
   LandingScreen: undefined
   NewWalletIntroScreen: undefined
@@ -52,9 +50,6 @@ type RootStackParamList = {
   NewContactScreen: undefined
   EditContactScreen: {
     contactId: string
-  }
-  SelectAddressScreen: {
-    nextScreen: PossibleNextScreenAfterDestination
   }
 }
